Revert "add useBlockParam/useBlockNumber hooks"
This reverts commit a377a719be3edc6455ae47567e477cdbf6a0792e.
