+++ dapp/src/components/BoardLink.tsx
@@ -1,11 +1,11 @@
 import { Board } from "dchan";
-import useBlockParam from "hooks/useBlockParam";
+import useBlockNumber from "hooks/useBlockNumber";
 import { Link } from "react-router-dom";
 import { Router } from "router";
 import IdLabel from "./IdLabel";
 
 export default function BoardLink({ board }: { board: Board }) {
-  const block = useBlockParam();
+  const block = useBlockNumber();
   return (
     <Link
       title={board.title}
