+++ dapp/src/components/post/PostHeader.tsx
@@ -13,7 +13,7 @@ import {
   unlockThread,
   unpinThread,
 } from "dchan/operations";
-import useBlockParam from "hooks/useBlockParam";
+import useBlockNumber from "hooks/useBlockNumber";
 import usePubSub from "hooks/usePubSub";
 import useSettings from "hooks/useSettings";
 import useUser from "hooks/useUser";
@@ -48,7 +48,7 @@ export default function PostHeader({
   const { isJannyOf } = useUser();
   const isOwner = accounts.length > 0 && accounts[0] === address;
   const [status, setStatus] = useState<string | object>();
-  const block = useBlockParam();
+  const block = useBlockNumber();
 
   const createdAt = fromBigInt(createdAtUnix);
   const relativeTime = createdAt.toRelative();
