+++ dapp/src/pages/idReference.tsx
@@ -12,10 +12,11 @@ import {
 } from "dchan";
 import SEARCH_BY_ID from "graphql/queries/search_by_id";
 import SEARCH_BY_ID_BLOCK from "graphql/queries/search_by_id_block";
-import useBlockNumber from "hooks/useBlockNumber";
 import { useEffect, useState } from "react";
 import { useHistory } from "react-router-dom";
 import { Router } from "router";
+import { parse as parseQueryString } from "query-string";
+import useBlockNumber from "hooks/useBlockNumber";
 
 interface IdSearchData {
   boardCreationEvent: BoardCreationEvent;
@@ -35,7 +36,14 @@ export default function IdReferencePage({ location, match: { params } }: any) {
   const history = useHistory();
 
   const id = `0x${params.id}`;
-  const queriedBlock = useBlockNumber();
+  const block = useBlockNumber();
+  let queriedBlock: number | undefined;
+  if (block) {
+    queriedBlock = parseInt(block);
+    if (isNaN(queriedBlock)) {
+      queriedBlock = undefined;
+    }
+  }
 
   console.log({ id, queriedBlock })
